ai-engine/requirements.txt
# Vision-AI Engine Dependencies
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
asyncpg>=0.29.0
aiosqlite>=0.19.0

# Redis
redis>=5.0.1

# AI/ML
torch>=2.6.0
torchvision>=0.21.0
ultralytics>=8.1.0
opencv-python-headless>=4.9.0.80
numpy>=1.26.3
Pillow>=10.2.0
scikit-learn>=1.4.0
scipy>=1.12.0
onnx>=1.15.0
onnxruntime>=1.17.0

# Image Processing
albumentations>=1.3.1
imagehash>=4.3.1

# MQTT
paho-mqtt>=1.6.1

# WebSocket
websockets>=12.0

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.9.3
aiofiles>=23.2.1
httpx>=0.26.0
orjson>=3.9.12
loguru>=0.7.2
apscheduler>=3.10.4
celery[redis]>=5.3.6

# Monitoring
prometheus-client>=0.19.0
psutil>=5.9.7

# Visualization & Reports
matplotlib>=3.8.2
seaborn>=0.13.1
reportlab>=4.0.8

# Email
aiosmtplib>=3.0.1
